Refactoring the import options view from the rolodex screen
